Extract some methods from base_helper to zones_helper to make it explicit they are only used in the zones admin pages
