+++ .rubocop_manual_todo.yml
@@ -389,6 +389,7 @@ Metrics/AbcSize:
   - app/helpers/order_cycles_helper.rb
   - app/helpers/spree/admin/navigation_helper.rb
   - app/helpers/spree/admin/base_helper.rb
+  - app/helpers/spree/admin/zones_helper.rb
   - app/helpers/spree/orders_helper.rb
   - app/jobs/subscription_placement_job.rb
   - app/mailers/producer_mailer.rb
