+++ .rubocop_todo.yml
@@ -332,6 +332,7 @@ Rails/OutputSafety:
     - 'app/controllers/spree/admin/reports_controller.rb'
     - 'app/helpers/angular_form_helper.rb'
     - 'app/helpers/spree/admin/base_helper.rb'
+    - 'app/helpers/spree/admin/zones_helper.rb'
     - 'app/helpers/spree/reports_helper.rb'
     - 'app/helpers/spree/admin/navigation_helper.rb'
     - 'app/serializers/api/product_serializer.rb'
